BA-2399 Code simplifications due to package updates
