+++ packages/design-system/components/native/inputs/SocialTextInput/index.tsx
@@ -3,12 +3,8 @@ import { forwardRef, useCallback } from 'react'
 import {
   NativeSyntheticEvent,
   TextInput as NativeTextInput,
-  Platform,
   TextInputContentSizeChangeEventData,
 } from 'react-native'
-import { NativeViewGestureHandler } from 'react-native-gesture-handler'
-import { TextInput as PaperTextInput } from 'react-native-paper'
-import { RenderProps } from 'react-native-paper/lib/typescript/components/TextInput/types'
 
 import { useTheme } from '../../../../providers/native'
 import { withNativeController } from '../../../../utils/native'
@@ -19,15 +15,7 @@ import type { SocialTextInputProps } from './types'
 
 const SocialTextInput = forwardRef<NativeTextInput, SocialTextInputProps>(
   (
-    {
-      children,
-      lineHeight = 22,
-      maxLines = 3,
-      contentStyle = {},
-      toolStyle = {},
-      shouldUseBottomSheetSafeComponents = false,
-      ...props
-    },
+    { children, lineHeight = 22, maxLines = 3, contentStyle = {}, toolStyle = {}, ...props },
     ref,
   ) => {
     const { isFocused, textHeight, setSocialInputState } = useSocialTextInput()
@@ -38,7 +26,6 @@ const SocialTextInput = forwardRef<NativeTextInput, SocialTextInputProps>(
       lineHeight,
       maxLines,
     })
-    const isAndroid = Platform.OS === 'android'
 
     const onContentSizeChange = useCallback(
       (event: NativeSyntheticEvent<TextInputContentSizeChangeEventData>) => {
@@ -46,23 +33,6 @@ const SocialTextInput = forwardRef<NativeTextInput, SocialTextInputProps>(
         setSocialInputState({ textHeight: Math.min(maxLines * lineHeight, componentHeight) })
       },
       [lineHeight, maxLines, setSocialInputState],
-    )
-
-    const renderTextInput = useCallback(
-      (renderProps: RenderProps) => {
-        const textInput = (
-          <NativeTextInput {...renderProps} onContentSizeChange={onContentSizeChange} />
-        )
-
-        if (shouldUseBottomSheetSafeComponents) {
-          return (
-            <NativeViewGestureHandler disallowInterruption>{textInput}</NativeViewGestureHandler>
-          )
-        }
-
-        return textInput
-      },
-      [onContentSizeChange, shouldUseBottomSheetSafeComponents],
     )
 
     const handleBlur = useCallback(
@@ -81,44 +51,20 @@ const SocialTextInput = forwardRef<NativeTextInput, SocialTextInputProps>(
       [props.onFocus, setSocialInputState],
     )
 
-    // Here, we treat android and ios separately: The lineHeight property of NativeTextInput does not work on android (but on iOS).
-    // As a workaround, we use PaperTextInput on android, for which this property works. But PaperTextInput sets a minimal height,
-    // which can only be overwritten by setting the height explicitly. This does not work on iOS, since onContentSizeChange is not invoked
-    // on iOS when a height is set.
-    const textInput = isAndroid ? (
-      <PaperTextInput
-        contentStyle={styles.contentStyle}
-        mode="outlined"
-        multiline
-        outlineStyle={styles.outlineStyle}
-        placeholder="Comment..."
-        placeholderTextColor={theme.colors.object.low}
-        ref={ref}
-        render={renderTextInput}
-        style={styles.input}
-        textColor={theme.colors.object.high}
-        {...props}
-        onBlur={handleBlur}
-        onFocus={handleFocus}
-      />
-    ) : (
-      <NativeTextInput
-        multiline
-        onContentSizeChange={onContentSizeChange}
-        placeholder="Comment..."
-        placeholderTextColor={theme.colors.object.low}
-        ref={ref}
-        style={styles.iosInput}
-        {...props}
-        onBlur={handleBlur}
-        onFocus={handleFocus}
-      />
-    )
-
     return (
       <View style={[contentStyle, styles.container]}>
         {/* TODO: Replies */}
-        {textInput}
+        <NativeTextInput
+          multiline
+          onContentSizeChange={onContentSizeChange}
+          placeholder="Comment..."
+          placeholderTextColor={theme.colors.object.low}
+          ref={ref}
+          style={styles.iosInput}
+          {...props}
+          onBlur={handleBlur}
+          onFocus={handleFocus}
+        />
         <View style={[toolStyle, styles.toolContainer]}>{children}</View>
       </View>
     )
