+++ packages/design-system/components/native/inputs/SocialTextInput/types.ts
@@ -12,5 +12,4 @@ export type SocialTextInputProps = PaperTextInputProps &
     maxLines?: number
     containerStyle?: StyleProp<ViewStyle>
     toolStyle?: StyleProp<ViewStyle>
-    shouldUseBottomSheetSafeComponents?: boolean
   }
